+++ packages/contracts/contracts/foundry/SphinxConstants.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.6.2 <0.9.0;
 import { NetworkInfo, NetworkType } from "./SphinxPluginTypes.sol";
 
 contract SphinxConstants {
-  string public constant sphinxLibraryVersion = 'v0.20.0';
+  string public constant sphinxLibraryVersion = 'v0.20.2';
   address public constant compatibilityFallbackHandlerAddress = 0xf48f2B2d2a534e402487b3ee7C18c33Aec0Fe5e4;
   address public constant multiSendAddress = 0xA238CBeb142c10Ef7Ad8442C6D1f9E89e07e7761;
   address public constant createCallAddress = 0x7cbB62EaA69F79e6873cD1ecB2392971036cFAa4;
@@ -14,7 +14,7 @@ contract SphinxConstants {
   address public constant safeSingletonAddress = 0xd9Db270c1B5E3Bd161E8c8503c55cEABeE709552;
   address public constant sphinxModuleImplAddress = 0x8f4E4d51B8050B0ff713eff1F88f3dD8b5e8a530;
 
-  uint8 internal constant numSupportedNetworks = 42;
+  uint8 internal constant numSupportedNetworks = 44;
 
   function getNetworkInfoArray() public pure returns (NetworkInfo[] memory) {
     NetworkInfo[] memory all = new NetworkInfo[](numSupportedNetworks);
@@ -270,6 +270,18 @@ contract SphinxConstants {
       chainId: 1918988905,
       networkType: NetworkType.Testnet
     });
+    all[42] = NetworkInfo({
+      network: Network.blast_sepolia,
+      name: "blast_sepolia",
+      chainId: 168587773,
+      networkType: NetworkType.Testnet
+    });
+    all[43] = NetworkInfo({
+      network: Network.blast,
+      name: "blast",
+      chainId: 81457,
+      networkType: NetworkType.Mainnet
+    });
     return all;
   }
 }
@@ -316,5 +328,7 @@ enum Network {
   zora,
   zora_sepolia,
   rari,
-  rari_sepolia
+  rari_sepolia,
+  blast_sepolia,
+  blast
 }