+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sphinx-labs/contracts
 
+## 0.20.2
+
+### Patch Changes
+
+- a6e1473: Filter calls to SphinxUtils out of state diff
+- caa8515: Remove storage gas cost subtraction on Moonbeam
+
+## 0.20.1
+
+### Patch Changes
+
+- fb439d6: Support installing Sphinx Library contracts via NPM
+
 ## 0.20.0
 
 ### Minor Changes